Move the reflection GUI into its application package
The problem was split across two packages, the driver in the solver
package and the GUI in the pilot, which left the next application no
single place to copy.  Move the GUI in beside the driver:
`_oblique.py` becomes `_mesh_sample.py`, `_field_render.py` keeps its
name, and `_solution_info.py` splits along the seam it already had
into `_panel.py` for the control widget and `_app.py` for the feature
that owns the solver, the viewer, and the timer.  `SolutionInfo`
becomes `ObliqueShockApp`, which says what it runs rather than where
it is shown.

`solvcon.pilot._euler` is gone, the application package binds its four
Qt names behind the pilot toggle, and the controller reaches the two
features through the package rather than their modules.  The tests
follow the modules they exercise, with only the import paths and the
feature name updated.
