+++ solvcon/pilot/apps/obsrefl/__init__.py
@@ -11,12 +11,27 @@ an incident oblique shock, the bottom slip wall reflects that shock, and the
 flow leaves through the non-reflective outflow on the right.
 """
 
+from ... import _pilot_core as _pcore
 from ._driver import ObliqueShock, ObliqueShockMesher, ObliqueShockRelation
+
+if _pcore.enable:
+    from ._app import ObliqueShockApp
+    from ._mesh_sample import ObliqueShockMesh
+    from ._panel import SolutionPanel
+else:
+    # Bind only the public names: a None module attribute would shadow the
+    # real submodule import in no-GUI builds.
+    ObliqueShockApp = None
+    ObliqueShockMesh = None
+    SolutionPanel = None
 
 __all__ = [
     'ObliqueShock',
+    'ObliqueShockApp',
+    'ObliqueShockMesh',
     'ObliqueShockMesher',
     'ObliqueShockRelation',
+    'SolutionPanel',
 ]
 
 # vim: set ff=unix fenc=utf8 et sw=4 ts=4 sts=4:
